refactor(frontend): / -> /Shared
