+++ packages/frontend/components/Shared/SizableTableCell.tsx
