+++ packages/frontend/components/Shared/Snackbar.tsx
@@ -12,8 +12,8 @@ import CloseIcon from "@mui/icons-material/Close"
 import CheckIcon from "@mui/icons-material/Check"
 import ErrorIcon from "@mui/icons-material/Error"
 import LaunchIcon from "@mui/icons-material/Launch"
-import { useSnack, Snack } from "../store/snackbar.store"
-import { transactionLink } from "../helpers/transactionInformations"
+import { useSnack, Snack } from "../../store/snackbar.store"
+import { transactionLink } from "../../helpers/transactionInformations"
 
 export function Snackbar() {
   const [snack] = useSnack((s) => s.notifications)
