+++ packages/frontend/pages/_app.tsx
@@ -6,7 +6,7 @@ import { ThemeProvider } from "@mui/material"
 
 import { theme } from "../styles/theme"
 import { useStore } from "../store"
-import { Snackbar } from "../components/Snackbar"
+import { Snackbar } from "../components/Shared/Snackbar"
 
 function MyApp({ Component, pageProps }: AppProps) {
   const init = useStore((state) => state.init)
